Fix: Move benchmark test from test/ to bench/ directory (PR #1112)
The test ActivateBufferReuseBenchmark.ts was causing CI failures with exit
code 143 (SIGTERM/killed) because:

1. It was a performance benchmark running in the unit test suite
2. It created large creatures via 450 repeated AddNeuron mutations
3. It ran 10,000+ iterations of activate() calls
4. When running in parallel with other tests, it exceeded time limits

This is not a memory leak - the test was simply inappropriate for the unit
test suite. Per project guidelines: "Unit tests are for functionality testing
only (not performance). Use benchmarks for performance validation."

Changes:
- Moved test/ActivateBufferReuseBenchmark.ts to bench/ActivateBufferReuse.ts
- Converted from Deno.test() to Deno.bench() API
- Simplified to use pre-created creatures (avoiding mutation overhead)
- Benchmark can now be run separately via: deno bench bench/ActivateBufferReuse.ts

The existing test/ActivateBufferReuse.ts remains for correctness testing.
